+++ shinobi_client/__init__.py
@@ -1,6 +1,7 @@
 from shinobi_client.client import ShinobiClient
 from shinobi_client.orms.user import ShinobiWrongPasswordError
 from shinobi_client._common import ShinobiSuperUserCredentialsRequiredError
+from shinobi_client.orms.monitor import ShinobiMonitorOrm, ShinobiMonitorAlreadyExistsError
 try:
     from shinobi_client.shinobi_controller import start_shinobi, ShinobiController
 except ImportError: