+++ shinobi_client/client.py
@@ -4,7 +4,9 @@ from dataclasses import dataclass
 @dataclass
 class ShinobiClient:
     """
-    Model of Shinobi installation.
+    Shinobi client.
+
+    Not thread safe.
     """
     host: str
     port: str
@@ -25,3 +27,7 @@ class ShinobiClient:
     def api_key(self) -> "ShinobiApiKey":
         from shinobi_client.api_key import ShinobiApiKey
         return ShinobiApiKey(self)
+
+    def monitor(self, email: str, password: str) -> "ShinobiMonitor":
+        from shinobi_client.orms.monitor import ShinobiMonitorOrm
+        return ShinobiMonitorOrm(self, email, password)