+++ shinobi_client/orms/user.py
@@ -2,31 +2,36 @@ import json
 from copy import deepcopy
 from dataclasses import dataclass
 
-from time import sleep
 from typing import Optional, Dict, Tuple
 
 import requests
-from logzero import logger
 
 from shinobi_client import ShinobiClient
-from shinobi_client._common import raise_if_errors, ShinobiSuperUserCredentialsRequiredError
+from shinobi_client._common import raise_if_errors, ShinobiSuperUserCredentialsRequiredError, wait_and_verify
 
 
 @dataclass
-class ShinobiWrongPasswordError(RuntimeError):
-    """
-    TODO
-    """
+class ShinobiWrongPasswordError(ValueError):
     email: str
     password: str
+
+
+@dataclass
+class ShinobiUserAlreadyExistsError(ValueError):
+    email: str
+
+
+@dataclass
+class ShinobiUserDoesNotExistError(ValueError):
+    email: str
 
 
 class ShinobiUserOrm:
     """
     Shinobi user ORM.
-    """
-    SUPPORTED_MODIFIABLE_PROPERTIES = {"password"}
 
+    Not thread safe.
+    """
     @staticmethod
     def _create_improved_user_entry(user: Dict) -> Dict:
         """
@@ -81,7 +86,7 @@ class ShinobiUserOrm:
         raise_if_errors(response, raise_if_json_not_ok=False)
 
         if not response.json()["ok"]:
-            # We can only assume this means that the password was incorrect...
+            # We can only assume this means that the password was incorrect/the user doesn't exist...
             raise ShinobiWrongPasswordError(email, password)
 
         user = response.json()["$user"]
@@ -111,17 +116,17 @@ class ShinobiUserOrm:
         raise_if_errors(response)
         return tuple(ShinobiUserOrm._create_improved_user_entry(user) for user in response.json()["users"])
 
-    def create(self, email: str, password: str, verify_create: bool = True) -> Dict:
+    def create(self, email: str, password: str, verify: bool = True) -> Dict:
         """
         Creates a user with the given details.
         :param email: email address of the user
         :param password: password for the user
-        :param verify_create: whether to wait to confirm that the user has been created
+        :param verify: whether to wait to confirm that the user has been created
         :return: details about created user
         """
         # Not trusting Shinobi's API to give back anything useful if the user already exists
         if self.get(email):
-            raise ValueError(f"User with email \"{email}\" already exists")
+            raise ShinobiUserAlreadyExistsError(email)
 
         # The required post does not align with the API documentation (https://shinobi.video/docs/api)
         # Exploiting the undocumented API successfully used by UI.
@@ -141,52 +146,51 @@ class ShinobiUserOrm:
         raise_if_errors(response)
         create_user = response.json()
 
-        # This is worth doing as Shinobi's API is all over the place - it happily returns OK for invalid requests
-        if verify_create:
-            if not self._wait_and_verify(email, True):
+        if verify:
+            if not wait_and_verify(lambda: self.get(email) is not None):
                 raise RuntimeError("Unable to verify created user")
 
         return ShinobiUserOrm._create_improved_user_entry(create_user["user"])
 
-    def modify(self, email: str, **kwargs) -> Optional[bool]:
+    def modify(self, email: str, *, password: str) -> bool:
         """
         Modify a user.
-        :param email: the email address of the user to modify.
-        :param kwargs: new property values (currently supported: `password`).
-        :return: whether the user was modified or `None` if not able to confirm
+        :param email: the email address of the user to modify
+        :param password: new password
+        :return: whether the user was modified
         """
-        unsupported_properties = set(kwargs.keys()) - {"mail"} - self.__class__.SUPPORTED_MODIFIABLE_PROPERTIES
-        if len(unsupported_properties) > 0:
-            raise NotImplementedError(f"Cannot modify user properties: {unsupported_properties}")
+        try:
+            user = self.get(email, password)
+            # The user can't be none as otherwise their credentials shouldn't have worked at getting the user's details
+            assert user is not None
+            return False
+        except ShinobiWrongPasswordError:
+            pass
 
-        existing_user = self.get(email)
-        if existing_user is None:
-            raise ValueError(f"Cannot modify user as they do not exist: {email}")
+        user = self.get(email)
+        if user is None:
+            raise ShinobiUserDoesNotExistError(email)
 
         data = {
             "mail": email,
-            "pass": kwargs["password"],
-            "password_again": kwargs["password"],
+            "pass": password,
+            "password_again": password,
         }
         account = {
             "mail": email,
-            "uid": existing_user["uid"],
-            "ke": existing_user["ke"]
+            "uid": user["uid"],
+            "ke": user["ke"]
         }
         response = requests.post(f"{self._base_url}/editAdmin", json=dict(data=data, account=account))
         raise_if_errors(response)
 
-        rows_changed = response.json().get("rowsChanged")
-        if rows_changed is None:
-            logger.info("Shinobi did not return information on whether the user has been changed")
-            return None
-        return rows_changed == 1
+        return True
 
-    def delete(self, email: str, verify_delete: bool = True) -> bool:
+    def delete(self, email: str, verify: bool = True) -> bool:
         """
         Deletes the user with the given email address.
         :param email: email address of user
-        :param verify_delete: whether to wait to confirm that the user has been deleted
+        :param verify: whether to wait to confirm that the user has been deleted
         :return: `True` if the user has been deleted, else `False` if they haven't because they didn't exist
         """
         user = self.get(email)
@@ -204,28 +208,8 @@ class ShinobiUserOrm:
         response = requests.post(f"{self._base_url}/deleteAdmin", json=dict(account=account))
         raise_if_errors(response)
 
-        if verify_delete:
-            if not self._wait_and_verify(email, False):
+        if verify:
+            if not wait_and_verify(lambda: self.get(email) is None):
                 raise RuntimeError(f"User with email \"{email}\" was not deleted")
 
         return True
-
-    def _wait_and_verify(self, email: str, user_should_exist: bool, *, wait_iterations: int = 10,
-                         iteration_wait_in_milliseconds_multiplier: int = 100):
-        """
-        Wait to verify if the user with the given email address exists (if they should) or doesn't (if they shouldn't).
-        :param email: email of user to check
-        :param user_should_exist: whether the user should exist
-        :param wait_iterations: number of times to try if state not valid
-        :param iteration_wait_in_milliseconds_multiplier: waiting for `wait_iterations` * `iteration_wait_in_milliseconds_multiplier`
-                                                          between each iteration
-        :return: `True` if verified state matches desired, else `False`
-        """
-        for i in range(wait_iterations):
-            user = self.get(email)
-            if user and user_should_exist:
-                return True
-            elif not user and not user_should_exist:
-                return True
-            sleep(iteration_wait_in_milliseconds_multiplier * i)
-        return False